refactor(go-server): extract function inventories
Keep query policy enforcement separate from the reviewed DuckDB name sets so extension bundles can grow without overwhelming the query package API.
