+++ packages/server/duckdb-server-go/pkg/functionset/functionset.go
@@ -1,4 +1,5 @@
-package query
+// Package functionset provides reviewed DuckDB function-name inventories for query policies.
+package functionset
 
 import (
 	"maps"
